% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{anscombe}
\alias{anscombe}
\title{anscombe transform of counts}
\usage{
anscombe(x)
}
\arguments{
\item{x}{vector of counrs}
}
\value{
return anscomb transform of x
}
\description{
anscombe transform of counts
}
\examples{
\dontrun{
x = anscombe(scd$getgene("gene_a"))
}
}
