% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ascb}
\alias{ascb}
\title{anscombe transform of counts}
\usage{
ascb(x, to_zero = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{vector of counrs}

\item{to_zero}{(bool default: TRUE) set ascb(x) to 0 when x = 0}

\item{na.rm}{remove NA values}
}
\value{
return anscomb transform of x
}
\description{
anscombe transform of counts
}
\examples{
\dontrun{
x = ascb(scd$getgene("gene_a"))
}
}
