% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{bca_loading}
\alias{bca_loading}
\title{compute bca loading}
\usage{
bca_loading(scd, by, norm_by, ncomp = 5, cells_l = FALSE,
  loading = F)
}
\arguments{
\item{by}{a factor defining the groups}

\item{norm_by}{a factor defining the groups with want to normalize}

\item{ncomp}{(default=5) number of component to compute}

\item{cells_l}{(default=FALSE) shall the cells loading be also returned}

\item{loading}{(default=FALSE) return only the genes contribution to the axes}

\item{x}{vector data.frame to vectorize}
}
\value{
return a matrix of cells coordinates or a list including this matrix
}
\description{
compute bca loading
}
\examples{
\dontrun{
x = bca_loading(scd, scd$getfeature("sex"))
}
}
