% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{binomial_fit}
\alias{binomial_fit}
\title{importFrom lme4 glmer}
\usage{
binomial_fit(data, formula, gene_name, family = "binomial", link = "log",
  zi = TRUE, v)
}
\description{
importFrom lme4 glmer
}
