% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{check_gene}
\alias{check_gene}
\title{display single or multiple genes as violin plot}
\usage{
check_gene(scd, gene, condition, file, transform = TRUE)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{gene}{gene name to display}

\item{condition}{condition to compare (must be features of scd)}

\item{file}{output file where to save the graph (optional)}

\item{transform}{(bool) use anscomb transform or raw counts}
}
\value{
a list() with the number of time each cells is classied as 'good' (
non is_blank looking) or 'bad' (blank looking)
}
\description{
display single or multiple genes as violin plot
}
\examples{
\dontrun{
check_gene(scd, "genes_a", "sexe")
}
}
