% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{classification for scRNASeq data}
\usage{
classification(scd, feature, ncores = 4, features, genes, algo = "spls_pls",
  selection = "stab", output_file = "", force = c(), v = F)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{feature}{the feature defining the groups to learn on}

\item{ncores}{(default:4) number of cpus to use}

\item{features}{a vector of the names of the feature to classifiy on}

\item{genes}{a vector of the names of the genes to classifiy on}

\item{algo}{is the algorithm to use}

\item{output_file}{if non empty file to save the classification results}

\item{force}{a list of features and genes}

\item{training}{is a vector of cell names to train on}
}
\value{
a list() with the result of the classification
}
\description{
classification for scRNASeq data
}
\examples{
\dontrun {
classif <- classification(scd, 500000)
}
}
