% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{DEA}
\alias{DEA}
\title{DEA for scRNASeq data}
\usage{
DEA(scd, formula_null, formula_full, b_cells, zi_threshold = 0.9, cpus = 4,
  v = F, folder_name)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{formula_null}{null model for the LR test}

\item{formula_full}{full model for the LR test}

\item{output_file}{if non empty file to save the classification results}
}
\value{
a list() with the result of the DEA
}
\description{
DEA for scRNASeq data
}
\examples{
\dontrun {

  b_cells <- scd$getfeature("QC_good") \%in\% T &
    !is.na(scd$getfeature("DEA_cell_type")) &
    scd$getfeature("day") \%in\% day
  mbatch_DEA_cell_type_DEA <- DEA(
    scd = scd,
    formula_null = "y ~ (1|batch)",
    formula_full = "y ~ (1|batch) + DEA_cell_type",
    b_cells = b_cells,
    cpus = 16,
    v = T,
    folder_name = paste0("results/cell_type/mbatch_", day, "_DEA_cell_type_DEA")
  )

}
}
