% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{DEA_LRT}
\alias{DEA_LRT}
\title{importFrom glmmADMB glmmadmb}
\usage{
DEA_LRT(models_result, gene_name, v, folder_name, model_family = "nbinom1")
}
\description{
importFrom glmmADMB glmmadmb
}
