% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{DEA_pbs}
\alias{DEA_pbs}
\title{helper function to run DEA in a script}
\usage{
DEA_pbs(Args = commandArgs())
}
\arguments{
\item{Args}{agument passsed when used within a script}

\item{scd_file}{RData file with an scTools data object 'data' saved in it}
}
\value{
write a paraload ready table with the parameters (as line) on which
to run every DEA() commands
}
\description{
helper function to run DEA in a script
}
\examples{
\dontrun{
DEA_pbs()
}
}
