% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dotplot}
\alias{dotplot}
\title{genes expression dotplot graph}
\usage{
dotplot(scd, title, clones_order, genes_order,
  rank_by = "pDEA_cell_type", file)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{title}{the title of the plot}

\item{clones_order}{vector of clone name to keep and to order by}

\item{genes_order}{vector of genes names to keep and to order by}

\item{rank_by}{name of the feature to order by}

\item{file}{path to save the plot (.pdf or .png)}
}
\description{
genes expression dotplot graph
}
\examples{
\dontrun{
dotplot(
  scd = scd,
  title = "cell type dotplot"
)
}
}
