% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ERCC}
\alias{ERCC}
\title{Select ERCC from gene list}
\usage{
ERCC(data, minus = FALSE)
}
\arguments{
\item{data}{scdata object}

\item{minus}{(default: false) if true return non ERCC}
}
\value{
return a vector of ERCC gene names
}
\description{
Select ERCC from gene list
}
\examples{
\dontrun{
genes = ERCC(data, minus = true)
data = data$select(genes = genes)
}
}
