% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{expressed}
\alias{expressed}
\title{return genes expressed with the following criteria}
\usage{
expressed(scd, zi_threshold = 0.9, alt = "lesser")
}
\arguments{
\item{scd}{scdata object}

\item{zi_threshold}{persentage of cells == 0}

\item{alt}{(default: lesser) should genes with lesser or greater number of
zeros be returned}
}
\value{
return list of genes names
}
\description{
return genes expressed with the following criteria
}
\examples{
\dontrun{
expressed_genes = expressed(scd, 0.90)
}
}
