% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{factorize}
\alias{factorize}
\title{factorize a vector or data.frame}
\usage{
factorize(x, columns)
}
\arguments{
\item{x}{vector data.frame to vectorize}

\item{columns}{number to vectorize in case of a data.frame}
}
\value{
return x with as.factor(as.vector(x)) applied
}
\description{
factorize a vector or data.frame
}
\examples{
\dontrun{
x = factorize(x)
}
}
