% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap_corr_cells}
\alias{heatmap_corr_cells}
\title{correlation heatmap of genes}
\usage{
heatmap_corr_cells(scd, features, factor = rep(TRUE, length(features)),
  genes_order, show_legend = TRUE, title = "", pca = FALSE,
  pCMF = FALSE, ncomp = 4, cpus = 4, dist_name = "manhattan",
  gene_size = 10, file)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{features}{features of the scd object to display}

\item{factor}{vector indicating which features should be dealt with like a
factor}

\item{genes_order}{genes order}

\item{show_legend}{(default: TRUE) should the legend be displayed}

\item{title}{title of the heatmap}

\item{ncomp}{(default = 5) number of metagene to compute (PCA)}

\item{file}{name of the pdf to save the heatmap}
}
\value{
return a heatmap object
}
\description{
correlation heatmap of genes
}
\examples{
\dontrun{
headmap_corr_genes(
  scd = scd,
  features = c("cell_type", "day"),
  cells_order = order(scd$getfeature("cell_type")),
  title = "cell type heatmap"
)
}
}
