% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap_corr_genes}
\alias{heatmap_corr_genes}
\title{correlation heatmap of cells}
\usage{
heatmap_corr_genes(scd, features, cells_order, factor = rep(TRUE,
  length(features)), show_legend = TRUE, title = "", pca = FALSE,
  simlr = FALSE, simlr_clust = 2, ncomp = 4, dist_name = "manhattan",
  file)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{features}{features of the scd object to display}

\item{cells_order}{cells order}

\item{factor}{vector indicating which features should be dealt with like a
factor}

\item{show_legend}{(default: TRUE) should the legend be displayed}

\item{title}{title of the heatmap}

\item{ncomp}{(default = 5) number of metagene to compute (PCA)}

\item{file}{name of the pdf to save the heatmap}
}
\value{
return a heatmap object
}
\description{
correlation heatmap of cells
}
\examples{
\dontrun{
headmap_corr_genes(
  scd = scd,
  features = c("cell_type", "day"),
  cells_order = order(scd$getfeature("cell_type")),
  title = "cell type heatmap"
)
}
}
