% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap_genes}
\alias{heatmap_genes}
\title{heatmap of genes}
\usage{
heatmap_genes(scd, features, cells_order = order(scd$getcells),
  genes_order = order(scd$getgenes), factor = rep(TRUE, length(features)),
  show_legend = TRUE, title = "", file, FUN = function(x) {     x <- log(x
  + 1)     x <- (x - mean(x))/sd(x) }, gene_size = round(scd$getngenes/12))
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{features}{features of the scd object to display}

\item{cells_order}{cells order}

\item{genes_order}{genes order}

\item{factor}{vector indicating which features should be dealt with like a
factor}

\item{show_legend}{(default: TRUE) should the legend be displayed}

\item{title}{title of the heatmap}

\item{file}{name of the pdf to save the heatmap}
}
\value{
return a heatmap object
}
\description{
heatmap of genes
}
\examples{
\dontrun{
headmap_genes(
  scd = scd,
  features = c("cell_type", "day"),
  cells_order = order(scd$getfeature("cell_type")),
  title = "cell type heatmap"
)
}
}
