% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load scRNASeq data}
\usage{
load_data(infos, counts)
}
\arguments{
\item{infos}{a text tabular information file on cells with cells in rows and
features in columns}

\item{counts}{a text tabular count file for cells in rows and genes in
columns}
}
\value{
a list with infos and counts associated and formated for the others
scRNASeq functions
}
\description{
load scRNASeq data
}
\examples{
\dontrun{
data <- load_data('data/infos.csv', 'data/counts.csv')
}
}
