% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_salmon}
\alias{load_data_salmon}
\title{load scRNASeq data}
\usage{
load_data_salmon(infos, counts, id_regexp = ".*_(P[0-9]{4}_[^_]+)_.*",
  ERCC_regexp = "^ERCC.*")
}
\arguments{
\item{infos}{a tabular csv file that contains information on the cells}

\item{counts}{path to kallisto quant.sf output}

\item{id_regexp}{a regular expression caputing the id of the cells in the
salmon folder output paths}

\item{ERCC_regexp}{a regular expression caputing genes names of the ERCC}

\item{...}{additional argument to be passed to read.table function}
}
\value{
scdata object
}
\description{
load scRNASeq data
}
\examples{
\dontrun{
data <- load_data_salmon('data/infos.csv', 'data/salmon_output')
}
}
