% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{order_2_groups}
\alias{order_2_groups}
\title{return order base on the rank FUN in groups}
\usage{
order_2_groups(scd, b_cells = NULL, cells = NULL, genes = NULL, by,
  FUN = function(x) {     mean(x)/sd(x) }, top = ncol(data),
  min_cell = 5)
}
\arguments{
\item{scd}{an scdata object}

\item{by}{factor to group on}

\item{FUN}{(default: mean) function on apply on factor in each groups}
}
\value{
return order
}
\description{
return order base on the rank FUN in groups
}
\examples{
\dontrun{
genes_order = order_by_groups(scd$getcounts, scd$get_feature("DEA_cell_type"))
}
}
