% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{order_by_factor}
\alias{order_by_factor}
\title{return order of genes by a covariate}
\usage{
order_by_factor(score, scd, tmp_file, top = FALSE, quant = 0.75, cpus = 4)
}
\arguments{
\item{score}{covariate to order on}

\item{scd}{scdata with genes to order}
}
\value{
return order
}
\description{
return order of genes by a covariate
}
\examples{
\dontrun{
genes_order = order_by_factor(scd$get_feature("pDEA_cell_type"), scd)
}
}
