% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{order_by_groups}
\alias{order_by_groups}
\title{return order base on the rank FUN in groups}
\usage{
order_by_groups(score, by, FUN = mean)
}
\arguments{
\item{score}{score to order on}

\item{by}{factor to group on}

\item{FUN}{(default: mean) function on apply on factor in each groups}
}
\value{
return order
}
\description{
return order base on the rank FUN in groups
}
\examples{
\dontrun{
cells_order = order_by_groups(scd$get_feature("pDEA_cell_type"), scd$get_feature("DEA_cell_type"))
}
}
