% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{order_TMP}
\alias{order_TMP}
\title{return order base on the rank FUN in groups}
\usage{
order_TMP(scd, by, FUN = mean, top = scd$getngenes)
}
\arguments{
\item{scd}{an scdata object}

\item{by}{factor to group on}

\item{FUN}{(default: mean) function on apply on factor in each groups}
}
\value{
return order
}
\description{
return order base on the rank FUN in groups
}
\examples{
\dontrun{
genes_order = order_TMP(scd$getcounts, scd$get_feature("DEA_cell_type"))
}
}
