% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pca_loading}
\alias{pca_loading}
\title{compute pca loading}
\usage{
pca_loading(scd, cells = FALSE, ncomp = 5)
}
\arguments{
\item{cells}{(default=FALSE) shall the cells loading be returned}

\item{x}{vector data.frame to vectorize}
}
\value{
return a matrix of cells (or genes) coordinates or a list including
this matrix
}
\description{
compute pca loading
}
\examples{
\dontrun{
x = pca_loading(scd)
}
}
