% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pca_plot_a_space}
\alias{pca_plot_a_space}
\title{pca of a dataset into the pca of another}
\usage{
pca_plot_a_space(data_a, data_b, color_a, color_b, main = "",
  name_a = "", name_b = "", file, axes = c(1, 2), size = 1)
}
\arguments{
\item{data_a}{first data set}

\item{data_b}{second dataset}

\item{color_a}{color of the first dataset}

\item{color_b}{color of the second dataset}

\item{main}{title of the plot}

\item{name_a}{name of the first data set}

\item{name_b}{name of the second data set}

\item{file}{name of the second data set}

\item{axes}{(default=c(1,2)) number of PCA axes to plot}

\item{size}{(default=1) size of the points}

\item{file}{path to save the plot (.pdf or .png)}
}
\description{
pca of a dataset into the pca of another
}
\examples{
\dontrun{
pca_plot_a_space(
  data_a = scd$select(b_cells = scd$getfeatures("cell_number") > 1),
  data_b = scd$select(b_cells = scd$getfeatures("cell_number") == 1),
  title = "bulk vs single cell"
)
}
}
