% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pCMF_loading}
\alias{pCMF_loading}
\title{compute pCMF loading}
\usage{
pCMF_loading(scd, cells = FALSE, ncomp = 5, cpus = 4, tmp_file)
}
\arguments{
\item{cells}{(default=FALSE) shall the cells loading be returned}

\item{x}{vector data.frame to vectorize}
}
\value{
return a matrix of cells (or genes) coordinates or a list including
this matrix
}
\description{
compute pCMF loading
}
\examples{
\dontrun{
x = pCMF_loading(scd)
}
}
