% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pCMF_plot}
\alias{pCMF_plot}
\title{plot the 2 first axis of a pCMF analysis}
\usage{
pCMF_plot(scd, color = NULL, shape = NULL, size = NULL,
  alpha = NULL, wrap = NULL, file, main = "", axes, res = FALSE,
  rainbow = FALSE, heatcolor = FALSE, is_contour, label = NULL,
  genes_list, arrow = FALSE, color_name = "day", return_data = FALSE,
  color_scale = NULL, tmp_file, ncomp = 2, ncores = 4)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{file}{output file where to save the graph (optional)}

\item{gene}{gene name to display}

\item{condition}{condition to compare (must be features of scd)}

\item{transform}{(bool) use anscomb transform or raw counts}
}
\value{
a list() with the number of time each cells is classied as 'good' (
non is_blank looking) or 'bad' (blank looking)
}
\description{
plot the 2 first axis of a pCMF analysis
}
\examples{
\dontrun{
check_gene(scd, "genes_a", "sexe")
}
}
