% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{per_genes_barplot}
\alias{per_genes_barplot}
\title{genes expression barplot graph}
\usage{
per_genes_barplot(scd, genes, features, order_by, color_by,
  quantile_cut = 0.95, main = "", file, decreasing = F)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{features}{features of the scd object to display}

\item{file}{name of the pdf to save the heatmap}

\item{decreasing}{(default: FALSE) should the reverse order of order_by be used ?}

\item{cells_order}{cells order}

\item{factor}{vector indicating which features should be dealt with like a
factor}

\item{show_legend}{(default: TRUE) should the legend be displayed}

\item{title}{title of the heatmap}
}
\value{
return a heatmap object
}
\description{
genes expression barplot graph
}
\examples{
\dontrun{
headmap_corr_genes(
  scd = scd,
  features = c("cell_type", "day"),
  cells_order = order(scd$getfeature("cell_type")),
  title = "cell type heatmap"
)
}
}
