% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{QC_boot}
\alias{QC_boot}
\title{QC for scRNASeq data}
\usage{
QC_boot(scd, iter,
  sample_size = length(which(scd$getfeature("cell_number") == 0)),
  is_blank = scd$getfeature("cell_number") == 0, output_file = "")
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{iter}{is the number of bootstrap sample to perform}

\item{sample_size}{of the cells sample to draw (the number of blanks for examples)}

\item{is_blank}{boolean vector set to TRUE if a cells is a blank}

\item{output_file}{if non empty file to save the classification results}
}
\value{
a list() with the number of time each cells is classied as 'good' (
non is_blank looking) or 'bad' (blank looking)
}
\description{
QC for scRNASeq data
}
\examples{
\dontrun{
classification <- QC_boot(scd, 500000)
}
}
