% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{QC_load_bootstraps}
\alias{QC_load_bootstraps}
\title{helper function for QC for scRNASeq data}
\usage{
QC_load_bootstraps(scd, paraload_folder)
}
\arguments{
\item{scd}{scdata object}

\item{paraload_folder}{dir where the QC_pbs output are saved}
}
\value{
return scdata object with a "QC_score" feature
}
\description{
helper function for QC for scRNASeq data
}
\examples{
\dontrun{
QC_paraload_parameters('results/QC/paraload_file.txt')
}
}
