% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{QC_paraload_parameters}
\alias{QC_paraload_parameters}
\title{helper function for QC for scRNASeq data}
\usage{
QC_paraload_parameters(paraload_file, bootstraps = 5e+05,
  job_boot_number = 50)
}
\arguments{
\item{paraload_file}{paraload table file}

\item{bootstraps}{total number of bootstrap to perform}

\item{job_boot_number}{of bootstrap per paraload jobs}
}
\value{
write a paraload ready table with the parameters (as line) on which
to run every QC_boot() commands
}
\description{
helper function for QC for scRNASeq data
}
\examples{
\dontrun{
QC_paraload_parameters('results/QC/paraload_file.txt')
}
}
