% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{QC_pbs}
\alias{QC_pbs}
\title{helper function to run QC in a script}
\usage{
QC_pbs(scd_file, Args = commandArgs())
}
\arguments{
\item{scd_file}{RData file with an scTools data object 'data' saved in it}

\item{Args}{agument passsed when used within a script}
}
\value{
write a paraload ready table with the parameters (as line) on which
to run every QC_boot() commands
}
\description{
helper function to run QC in a script
}
\examples{
\dontrun{
QC_paraload_parameters('results/QC/paraload_file.txt')
}
}
