% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{random_sample_data}
\alias{random_sample_data}
\title{random sample of scRNASeq data}
\usage{
random_sample_data(data, number, replace = FALSE)
}
\arguments{
\item{data}{a scdata object}

\item{number}{number of cells to sample}

\item{replace}{should sampling be with replacement?}
}
\value{
a scdata object with number of cells
}
\description{
random sample of scRNASeq data
}
\examples{
\dontrun{
rdata <- random_sample_data(data, 150)
}
}
