% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{split_heatmap}
\alias{split_heatmap}
\title{correlation heatmap of genes split in two by zi}
\usage{
split_heatmap(scd, features, factor = rep(TRUE, length(features)), genes,
  genes_order, cells_order, zi_split = 0.5, max_genes = 100,
  show_legend = TRUE, title = "", pca = FALSE, pCMF = FALSE,
  ncomp = 4, cpus = 4, dist_name = "manhattan", gene_size = 3,
  file)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{features}{features of the scd object to display}

\item{factor}{vector indicating which features should be dealt with like a
factor}

\item{genes_order}{genes order}

\item{show_legend}{(default: TRUE) should the legend be displayed}

\item{title}{title of the heatmap}

\item{ncomp}{(default = 5) number of metagene to compute (PCA)}

\item{file}{name of the pdf to save the heatmap}
}
\value{
return a heatmap object
}
\description{
correlation heatmap of genes split in two by zi
}
\examples{
\dontrun{
split_heatmap(
  scd = scd,
  features = c("cell_type", "day"),
  genes = DEG_list,
  title = "cell type heatmap"
)
}
}
