% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{top_2_groups}
\alias{top_2_groups}
\title{return top and bottom of list for order base on the rank FUN in groups}
\usage{
top_2_groups(scd, b_cells = NULL, top = 100, cells = NULL,
  genes = NULL, by, FUN = function(x) {     mean(x[x != 0]) *
  length(x[x != 0])/length(x)/sd(x[x != 0]) })
}
\arguments{
\item{scd}{an scdata object}

\item{top}{number of item to return}

\item{by}{factor to group on}

\item{FUN}{(default: mean) function on apply on factor in each groups}
}
\value{
return indice of top and bottom item
}
\description{
return top and bottom of list for order base on the rank FUN in groups
}
\examples{
\dontrun{
genes_top = top_by_groups(scd$getcounts, scd$get_feature("DEA_cell_type"), 100)
}
}
