% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{TPM}
\alias{TPM}
\title{compute TPM}
\usage{
TPM(counts, len)
}
\arguments{
\item{counts}{matrix or vector}

\item{len}{genes length vector}
}
\value{
return TPM
}
\description{
compute TPM
}
\examples{
\dontrun{
x = TPM(scd$getcounts, len)
}
}
