% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{TPM_matrix}
\alias{TPM_matrix}
\title{compute TPM}
\usage{
TPM_matrix(counts, len)
}
\arguments{
\item{counts}{matrix}

\item{len}{genes length vector}
}
\value{
return TPM
}
\description{
compute TPM
}
\examples{
\dontrun{
x = TPM_matrix(scd$getcounts, len)
}
}
