% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{TPM_vector}
\alias{TPM_vector}
\title{compute TPM}
\usage{
TPM_vector(counts, len)
}
\arguments{
\item{counts}{counts vector}

\item{len}{genes length vector}
}
\value{
return TPM
}
\description{
compute TPM
}
\examples{
\dontrun{
x = TPM_vector(scd$getcounts[1, ], len)
}
}
