% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{vectorize}
\alias{vectorize}
\title{vectorize a vector or data.frame}
\usage{
vectorize(x, columns)
}
\arguments{
\item{x}{vector data.frame to vectorize}

\item{columns}{number to vectorize in case of a data.frame}
}
\value{
return x with as.numeric(as.vector(x)) applied
}
\description{
vectorize a vector or data.frame
}
\examples{
\dontrun{
x = vectorize(x)
}
}
