% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{wca_norm}
\alias{wca_norm}
\title{compute wca normalized pca}
\usage{
wca_norm(scd, by, ncomp = 2, keep_zero = FALSE)
}
\arguments{
\item{by}{a factor defining the groups}

\item{keep_zero}{(default=FALSE) should the zero values be keept}

\item{x}{vector data.frame to vectorize}
}
\value{
return cells coordinates normalized for 'by' variance
}
\description{
compute wca normalized pca
}
\examples{
\dontrun{
x = wca_loading(scd, scd$getfeature("sex"))
}
}
