% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{winsorize}
\alias{winsorize}
\title{compute winsorize transform of the data}
\usage{
winsorize(x, up_quantile = 0.99, down_quantile, genes_as_col = T)
}
\arguments{
\item{x}{matrix, data.frame or vector of counts}

\item{up_quantile}{(default:0.99) upper quantile to winsorize the data
(per genes if 2 dimentional)}

\item{down_quantile}{if set, lower quantile to winsorize the data
(per genes if 2 dimentional)}

\item{genes_as_col}{(default: T) if a matrix of data.frame, are the genes
in column ?}
}
\value{
return winsorized x
}
\description{
compute winsorize transform of the data
}
\examples{
\dontrun{
x = winsorize(scd$getcounts, up_quantile = 0.90)
}
}
