% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{ziNB_fit}
\alias{ziNB_fit}
\title{importFrom glmmADMB glmmadmb}
\usage{
ziNB_fit(data, formula, gene_name, family = "nbinom1", link = "log", v)
}
\description{
importFrom glmmADMB glmmadmb
}
