% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{zinorm}
\alias{zinorm}
\title{return scalled counts for zidata}
\usage{
zinorm(scd, cpus = 4, v = F, file, zi_scale = TRUE,
  sd_scale = TRUE)
}
\arguments{
\item{scd}{scdata object to scale}
}
\value{
return scdata object with scaled counts
}
\description{
return scalled counts for zidata
}
\examples{
\dontrun{
scd_norm = zinorm(scd)
}
}
