% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{classification for scRNASeq data}
\usage{
classification(scd, feature, training, ncores = 4, algo = "mspls_mpls",
  output_file = "")
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{feature}{is the name of the feature to classifiy on}

\item{training}{is a vector of cell names to train on}

\item{algo}{is the algorithm to use}

\item{output_file}{if non empty file to save the classification results}
}
\value{
a list() with the result of the classification
}
\description{
classification for scRNASeq data
}
\examples{
\dontrun {
classif <- classification(scd, 500000)
}
}
