% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{DEA}
\alias{DEA}
\title{DEA for scRNASeq data}
\usage{
DEA(scd, formula_null, formula_full, b_cells, zi_threshold = 0.9, cpus = 4,
  v = F, folder_name)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{formula_null}{null model for the LR test}

\item{formula_full}{full model for the LR test}

\item{output_file}{if non empty file to save the classification results}
}
\value{
a list() with the result of the DEA
}
\description{
DEA for scRNASeq data
}
\examples{
\dontrun {
classif <- classification(scd, 500000)
}
}
