% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{DEA}
\alias{DEA}
\title{DEA for scRNASeq data}
\usage{
DEA(scd, formula_null, formula_full, b_cells, zi_threshold = 0.9,
  family = "nbinom1", cpus = 4, v = F, folder_name, zero_threshold = 5,
  continuous = c())
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{formula_null}{null model for the LR test}

\item{formula_full}{full model for the LR test}

\item{b_cells}{boolean vector of cells to keep in scd}

\item{zi_threshold}{(default: 0.9) work with genes with less that 0.9 zeros}

\item{family}{(default: "nbinom1") model to use for the DEA in
c("nbinom1", "binomial")}

\item{cpus}{(default: 4) number of cpus to use}

\item{v}{(default: F) print stuff}

\item{folder_name}{folder to write down intermendiate DEA results}

\item{zero_threshold}{(default: 5) counts lesser than this threshold will be
considered as zeros for the binomial regression}

\item{continuous}{vector of feature names that should be keep continuous (not
converted into factors}
}
\value{
a list() with the result of the DEA
}
\description{
DEA for scRNASeq data
}
\examples{
\dontrun {

  b_cells <- scd$getfeature("QC_good") \%in\% T &
    !is.na(scd$getfeature("DEA_cell_type")) &
    scd$getfeature("day") \%in\% day
  mbatch_DEA_cell_type_DEA <- DEA(
    scd = scd,
    formula_null = "y ~ (1|batch)",
    formula_full = "y ~ (1|batch) + DEA_cell_type",
    b_cells = b_cells,
    cpus = 16,
    v = T,
    folder_name = paste0("results/cell_type/mbatch_", day, "_DEA_cell_type_DEA")
  )

}
}
