% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{gene_cov}
\alias{gene_cov}
\title{classification for scRNASeq data}
\usage{
gene_cov(scd, gene, cpus = 4, tmp_file)
}
\arguments{
\item{scd}{is a scRNASeq data object}

\item{gene}{gene_name to find}

\item{cpus}{(default:4) number of cpus to use}
}
\value{
a vector with the name of the genes selected
}
\description{
classification for scRNASeq data
}
\examples{
\dontrun {
gene_list <- gene_cov(scd, "LTS")
}
}
