% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load scRNASeq data}
\usage{
load_data(infos, counts, regexp = ".*", ...)
}
\arguments{
\item{infos}{a text tabular information file on cells with cells in rows and
features in columns}

\item{counts}{a text tabular count file for cells in rows and genes in
columns. Can also be a folder in which case every files in the subtree will
be loaded'}

\item{regexp}{a regular expression marching the file we want to load in case
of counts is a folder}

\item{...}{additional argument to be passed to read.table function}
}
\value{
a list with infos and counts associated and formated for the others
scRNASeq functions
}
\description{
load scRNASeq data
}
\examples{
\dontrun{
data <- load_data('data/infos.csv', 'data/counts.csv')
}
}
