% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_salmon}
\alias{load_data_salmon}
\title{load_data_salmon scRNASeq data}
\usage{
load_data_salmon(infos, counts,
  id_regexp = ".*_(P[0-9]{4}_[0-9]{1,4})_.*",
  id_regexp_b = "P[0-9]{4}_[0-9]{1,4}", ERCC_regexp = "^ERCC.*",
  feature = "counts", infos_sep = ",", tximport_obj = "", ...)
}
\arguments{
\item{infos}{a tabular csv file that contains information on the cells}

\item{counts}{path to kallisto quant.sf output}

\item{id_regexp}{a regular expression caputing the id of the cells in the
salmon folder output paths}

\item{id_regexp_b}{a regular expression match the path of the quant.sf files}

\item{ERCC_regexp}{a regular expression caputing genes names of the ERCC}

\item{feature}{the feature to extract from salmon output}

\item{...}{additional argument to be passed to read.table function}
}
\value{
scdata object
}
\description{
load_data_salmon scRNASeq data
}
\examples{
\dontrun{
data <- load_data_salmon('data/infos.csv', 'data/salmon_output')
}
}
