% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize}
\alias{normalize}
\title{normalize counts in a scdata object}
\usage{
normalize(scd, b_cells = scd$getfeature("QC_good") \%in\% T,
  method = "SCnorm", cpus = 4, tmp_file, v = F)
}
\arguments{
\item{scd}{scdata object}

\item{method}{(default = "SCnorm") to use for the normalization}

\item{cpus}{(default = 4) number of cpus to use (if possible)}

\item{v}{(default = FALSE) verbose mode}
}
\value{
return scdata object
}
\description{
normalize counts in a scdata object
}
\examples{
\dontrun{
scd_norm <- normalize(scd)
}
}
