% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{scdata}
\alias{scdata}
\title{scRNASeq data object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
scdata
}
\arguments{
\item{infos}{a data.frame of cells (rows) features (columns)}

\item{counts}{a matrix of counts with cells as rows and genes as columns}
}
\value{
an scdata R6 R object
}
\description{
scRNASeq data object
}
\examples{
data <- scrna$new(infos, counts)
}
\keyword{datasets}
