% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{class}
\name{scdata}
\alias{scdata}
\title{scRNASeq data object}
\format{An \code{\link{R6Class}} generator object}
\usage{
scdata
}
\description{
scRNASeq data object
}
\keyword{counts}
\keyword{infos}
