% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{ziNB_fit}
\alias{ziNB_fit}
\title{importFrom glmmADMB glmmadmb
importFrom MASS glm.nb}
\usage{
ziNB_fit(data, formula, gene_name, model_family = "nbinom1",
  link = "log", zi = TRUE, v)
}
\description{
importFrom glmmADMB glmmadmb
importFrom MASS glm.nb
}
